Add support for partially instrumenting PG gem
This patch adds support for sync_exec, sync_exec_params, async_exec, and
async_exec_params. There are more methods that could be traced, and this
patch should have unit tests, but this works for Braze currently since
the patch is so simple and we don't really use methods beyond these.